refactor: split websocket package into lobby/shared and rename GameController to LobbySocketController
