+++ src/test/java/com/codenames/codenames/backend/lobby/api/dto/JoinMessageTest.java
@@ -1,4 +1,4 @@
-package com.codenames.codenames.backend.websocket;
+package com.codenames.codenames.backend.lobby.api.dto;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
